requirements.txt
Flask
Flask-SocketIO
Werkzeug
google-generativeai
python-dotenv
requests